changed/simplified: - ru.otus - removed @Table (where isn't needed) - changed fields to readable-ones in sql table and removed @Column
