--- buyer/src/main/java/com/example/buyer/BuyerApplication.java
@@ -1,13 +1,0 @@
-package com.example.buyer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BuyerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BuyerApplication.class, args);
-    }
-
-}
