+++ buyer/src/main/java/ru/otus/buyer/BuyerApplication.java
@@ -1,0 +1,13 @@
+package ru.otus.buyer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BuyerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BuyerApplication.class, args);
+    }
+
+}
