--- searcher/src/main/java/com/example/searcher/configuration/AppConfig.java
@@ -1,16 +1,0 @@
-package com.example.searcher.configuration;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class AppConfig {
-
-    @Bean
-    RestTemplate restTemplate() {
-
-        return new RestTemplate();
-    }
-}
