--- searcher/src/main/java/com/example/searcher/configuration/TravelPayoutProperties.java
@@ -1,16 +1,0 @@
-package com.example.searcher.configuration;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Getter
-@Setter
-@ConfigurationProperties(prefix = "travelpayout")
-
-public class TravelPayoutProperties {
-    private String token;
-    private String url;
-    private String searchUrl;
-    private String countriesUrl;
-}
