--- searcher/src/main/java/com/example/searcher/entity/Country.java
@@ -1,27 +1,0 @@
-package com.example.searcher.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-@Table(name = "country")
-@Getter
-@Setter
-public class Country {
-
-    @Id
-    @Column(name = "code")
-    private String id;
-
-    @Column(name = "name")
-    private String name;
-
-    @OneToMany(mappedBy = "country", fetch = FetchType.LAZY)
-    private List<City> cities;
-
-    @OneToMany(mappedBy = "country", fetch = FetchType.LAZY)
-    private List<Airport> airports;
-}
