--- searcher/src/main/java/com/example/searcher/entity/TestTable.java
@@ -1,25 +1,0 @@
-package com.example.searcher.entity;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class TestTable {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    Integer id;
-
-    String name;
-
-    Long someNumber;
-}
