--- searcher/src/main/java/com/example/searcher/model/TicketSearchResult.java
@@ -1,16 +1,0 @@
-package com.example.searcher.model;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class TicketSearchResult {
-    private boolean success;
-
-    @JsonFormat(with = JsonFormat.Feature.ACCEPT_SINGLE_VALUE_AS_ARRAY)
-    private List<Ticket> data;
-
-
-}
