--- searcher/src/main/java/com/example/searcher/repository/AirportRepository.java
@@ -1,9 +1,0 @@
-package com.example.searcher.repository;
-
-import com.example.searcher.entity.Airport;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AirportRepository extends JpaRepository<Airport, Long> {
-}
