--- searcher/src/main/java/com/example/searcher/repository/CityRepository.java
@@ -1,9 +1,0 @@
-package com.example.searcher.repository;
-
-import com.example.searcher.entity.City;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CityRepository extends JpaRepository<City, Long> {
-}
