--- searcher/src/main/java/com/example/searcher/repository/TicketRepository.java
@@ -1,9 +1,0 @@
-package com.example.searcher.repository;
-
-import com.example.searcher.entity.Ticket;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
-}
