--- searcher/src/main/java/com/example/searcher/service/TicketListService.java
@@ -1,8 +1,0 @@
-package com.example.searcher.service;
-
-import common_dto.SearchRequestDto;
-import common_dto.SearchResultDtoList;
-
-public interface TicketListService {
-    SearchResultDtoList getDtoTicketList(SearchRequestDto dto);
-}
