+++ searcher/src/main/java/ru/otus/searcher/configuration/AppConfig.java
@@ -1,0 +1,16 @@
+package ru.otus.searcher.configuration;
+
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    RestTemplate restTemplate() {
+
+        return new RestTemplate();
+    }
+}
