+++ searcher/src/main/java/ru/otus/searcher/configuration/TravelPayoutProperties.java
@@ -1,0 +1,16 @@
+package ru.otus.searcher.configuration;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "travelpayout")
+
+public class TravelPayoutProperties {
+    private String token;
+    private String url;
+    private String searchUrl;
+    private String countriesUrl;
+}
