+++ searcher/src/main/java/ru/otus/searcher/controllers/CountrySearchController.java
@@ -1,0 +1,23 @@
+package ru.otus.searcher.controllers;
+
+import ru.otus.searcher.service.CountrySearchService;
+import common_dto.CountryDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+
+@RestController
+@RequestMapping("api/country-search")
+@RequiredArgsConstructor
+public class CountrySearchController {
+
+    private final CountrySearchService countrySearchService;
+
+    @GetMapping
+    public List<CountryDto> getCountries() {
+
+        return countrySearchService.getCountry();
+    }
+}
