+++ searcher/src/main/java/ru/otus/searcher/converter/TicketSearchResultToSearchResultDTOConverter.java
@@ -1,0 +1,43 @@
+package ru.otus.searcher.converter;
+
+import ru.otus.searcher.model.Ticket;
+import ru.otus.searcher.model.TicketSearchResult;
+import common_dto.SearchResultDto;
+import common_dto.SearchResultDtoList;
+import lombok.experimental.UtilityClass;
+
+import java.util.*;
+
+@UtilityClass
+public class TicketSearchResultToSearchResultDTOConverter {
+
+//    TicketSearchResultToSearchResultDTOConverter
+//            .convert(Objects
+//            .requireNonNull(response
+//            .getBody()));
+
+    public SearchResultDtoList convert(TicketSearchResult result) {
+
+        List<SearchResultDto> searchResultDtoList = new ArrayList<>();
+
+        for (Ticket actualTicket : result.getData()) {
+
+            SearchResultDto resultDto = SearchResultDto
+                    .builder()
+                    .departCity(actualTicket.getOrigin())
+                    .arriveCity(actualTicket.getDestination())
+                    .site(actualTicket.getGate())
+                    .price(actualTicket.getValue())
+                    //.localDate(actualTicket.getFoundAt())
+                    .departDate(actualTicket.getDepartDate())
+                    .returnDate(actualTicket.getReturnDate())
+                    .numberOfChanges(actualTicket.getNumberOfChanges())
+                    .duration(actualTicket.getDuration())
+                    .distance(actualTicket.getDistance())
+                    .build();
+            searchResultDtoList.add(resultDto);
+        }
+
+        return new SearchResultDtoList(searchResultDtoList);
+    }
+}
