+++ searcher/src/main/java/ru/otus/searcher/entity/Airport.java
@@ -1,0 +1,27 @@
+package ru.otus.searcher.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+public class Airport {
+
+    @Id
+    private String code;
+
+    private String typeOfTransport;
+
+    private String name;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name =  "country_code")
+    private Country country;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name =  "city_code")
+    private City city;
+}
