+++ searcher/src/main/java/ru/otus/searcher/model/TicketSearchResult.java
@@ -1,0 +1,16 @@
+package ru.otus.searcher.model;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class TicketSearchResult {
+    private boolean success;
+
+    @JsonFormat(with = JsonFormat.Feature.ACCEPT_SINGLE_VALUE_AS_ARRAY)
+    private List<Ticket> data;
+
+
+}
