+++ searcher/src/main/java/ru/otus/searcher/repository/AirportRepository.java
@@ -1,0 +1,9 @@
+package ru.otus.searcher.repository;
+
+import ru.otus.searcher.entity.Airport;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AirportRepository extends JpaRepository<Airport, Long> {
+}
