+++ searcher/src/main/java/ru/otus/searcher/repository/CountryRepository.java
@@ -1,0 +1,9 @@
+package ru.otus.searcher.repository;
+
+import ru.otus.searcher.entity.Country;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CountryRepository extends JpaRepository<Country, Long> {
+}
