+++ searcher/src/main/java/ru/otus/searcher/repository/TicketRepository.java
@@ -1,0 +1,9 @@
+package ru.otus.searcher.repository;
+
+import ru.otus.searcher.entity.Ticket;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
+}
