+++ searcher/src/main/java/ru/otus/searcher/service/CountrySearchService.java
@@ -1,0 +1,30 @@
+package ru.otus.searcher.service;
+
+import ru.otus.searcher.configuration.TravelPayoutProperties;
+import common_dto.CountryDto;
+import lombok.NonNull;
+import org.apache.http.client.utils.URIBuilder;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.List;
+
+@Service
+public class CountrySearchService {
+    private final RestTemplate restTemplate;
+    private final URIBuilder builder;
+
+     public CountrySearchService(RestTemplate restTemplate, TravelPayoutProperties travelPayoutProperties){
+        this.restTemplate = restTemplate;
+        this.builder = new URIBuilder()
+                .setScheme("https")
+                .setHost(travelPayoutProperties.getUrl())
+                .setPath(travelPayoutProperties.getCountriesUrl());
+    }
+
+    public List<CountryDto> getCountry(){
+        @NonNull
+        CountryDto[] responseArray = restTemplate.getForEntity(builder.toString(),CountryDto[].class).getBody();
+        return List.of(responseArray);
+    }
+}
