+++ searcher/src/main/java/ru/otus/searcher/service/TicketListService.java
@@ -1,0 +1,8 @@
+package ru.otus.searcher.service;
+
+import common_dto.SearchRequestDto;
+import common_dto.SearchResultDtoList;
+
+public interface TicketListService {
+    SearchResultDtoList getDtoTicketList(SearchRequestDto dto);
+}
