+++ searcher/src/main/resources/application.yml
@@ -12,8 +12,6 @@ spring:
     url: jdbc:postgresql://localhost:5432/postgres
     username: postgres
     password: postgres
-    #dbcp2:
-      #default-catalog: postgres
   jpa:
     properties:
       hibernate:
