+++ searcher/src/main/resources/db/migration/V1.0.2__FlightSearcher.sql
@@ -9,13 +9,12 @@ create table if not exists city
     code         varchar(16) primary key,
     name         varchar(64),
     country_code varchar(16) references country (code)
-
 );
 
 create table if not exists ticket
 (
     id          serial primary key,
-    value       bigint,
+    price       bigint,
     origin      varchar(16) references city (code),
     destination varchar(16) references city (code)
 );
@@ -23,7 +22,7 @@ create table if not exists ticket
 create table if not exists airport
 (
     code        varchar(16) primary key,
-    iata_type   varchar(16),
+    type_of_transport   varchar(16),
     name   varchar(64),
     country_code varchar(16) references city (code),
     city_code varchar(16) references city (code)
